Adds XOAUTH2 support, intended for Gmail.
Makes --authmechX accept "XOAUTH2", which then uses the password for parameters.

--passwordX needs to be in the following format:
"<service account>;<keyfile location>;<keyfile password>"

<service account> is the name of the Google Developer API service account.

<keyfile location> is the location of the keyfile associated with it.

<keyfile password> is the password to access the keyfile. Entering this
password is optional--it will assume "notasecret" if not provided, which
is the default password Google uses with the keyfiles.

Example arguments for using XOAUTH might look like this:

--authmech1 xoauth2 --password1 '123456789012-abcd0abcde0ab1abc2345ab6abcdefgh@developer.gserviceaccount.com;some_keyfile.p12'

This assumes "some_keyfile" is in the current directory, and uses the default
"notasecret" password.