ipv6 and gmail
Something is not working in the current implementation with gmail. I think it might be related to ipv6, as forcing imapsync to use ipv4 seems to work.

The problem is reported over and over in newsgroups. While I'm not convinced that this is a good solution, it is at least a workaround.