XOAUTH2 support for Google JSON Data
When creating a service account, Google has switched to providing
a .json file with the private key, oauth client id, and other data
instead of an encrypted .p12 file with the just the private key.

This adds support to read one of these .json files instead of the
.p12 format. .p12 files are still supported, and which format is
read is determined by the formatting of the argument.

This is more convenient, but also avoids confusion around the
oauth client ids. Previously, the client id matches the service
account name Google generated. Now they are different, but this
fact isn't terribly apparent from their management interface.

Also added support for spaces in the filename for the .p12 files.