Give the user the option to not port some IMAP keywords to GMail labels
In my own conversion from Zimbra to GMail there were several non-user
keywords that I assume Zimbra was using internally such as $Forwarded,
Forwarded (yes, both with and without the $), and $MDNSent.  Following
the folder exclude model, add a --excludetags option that accepts a
regex of tags (IMAP keywords) to not convert to GMail labels.